Set r, g and b to int to remove typing issue and respect the intent of the doc-string.

+++ adafruit_character_lcd/character_lcd.py
@@ -715,9 +715,9 @@ class Character_LCD_RGB(Character_LCD):
             if color >> 24:
                 raise ValueError("Integer color value must be positive and 24 bits max")
             # NOTE: convert to 0-100
-            r = (color >> 16) / 2.55
-            g = ((color >> 8) & 0xFF) / 2.55
-            b = (color & 0xFF) / 2.55
+            r = int((color >> 16) / 2.55)
+            g = int(((color >> 8) & 0xFF) / 2.55)
+            b = int((color & 0xFF) / 2.55)
             color = [r, g, b]  # This is List[float], should be List[int]
         self._color = color
         for number, pin in enumerate(self.rgb_led):
